IIGRL-main/requirements.txt
h5py==2.10.0
idna==3.4
importlib-metadata==4.8.3
joblib==1.1.1
Keras-Preprocessing==1.1.2
Markdown==3.3.7
memory-profiler==0.61.0
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-version==0.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.23.2
scipy==1.4.1
six==1.16.0
sklearn==0.0.post1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.15
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.6.0
